app/requirements.txt
fastapi
uvicorn
python-multipart
anthropic
supabase
python-dotenv
pydantic
websockets==11.0.3
